refactor: a core subpackage of shared leaves, builders with what they build
Group the loose top-level modules into a layered structure instead of the flat
_params/_factory and the type aliases sitting in the package root:

- core/ holds only the shared leaves the whole tree imports — the type
  vocabulary (_types), the shared constants (_params), and the shared building
  primitives (_errors.BuildError, _reference._load_object). core/__init__
  re-exports the types, the params and BuildError; every module imports the
  shared vocabulary from the public core, not from a private submodule.
- Each builder lives in the package that owns what it builds and is re-exported
  from there: build_dataloader in dataloaders, build_bettor in evaluation,
  build_venue in execution. Every import runs downward (core -> domain packages
  -> surfaces), so no import cycle can form.
- The top-level sportsbet/__init__ is left holding only its front-page docstring;
  it no longer re-exports a subpackage's surface a second time.
- Deduplicate STATUSES, which the factory redefined verbatim from core._params.

Constitution 1.5.0 -> 1.6.0: the top level holds subpackages and its __init__,
not loose modules; the shared leaves live in core; a builder lives with what it
builds; a public name is re-exported once, at its owning surface; the top package
__init__ may carry a fuller front-page docstring.

Full gate green on 3.11/3.12/3.13 and the docs build.
