+++ src/sportsbet/dataloaders/__init__.py
@@ -3,10 +3,12 @@
 from __future__ import annotations
 
 from ._base import BaseDataLoader, load_dataloader
+from ._factory import build_dataloader
 from ._sourced import DataLoader
 
 __all__: list[str] = [
     'BaseDataLoader',
     'DataLoader',
+    'build_dataloader',
     'load_dataloader',
 ]
