+++ src/sportsbet/evaluation/__init__.py
@@ -11,6 +11,7 @@ from ._base import (
     save_bettor,
 )
 from ._classifier import ClassifierBettor
+from ._factory import build_bettor
 from ._model_selection import BettorGridSearchCV, backtest
 from ._rules import OddsComparisonBettor
 
@@ -20,6 +21,7 @@ __all__: list[str] = [
     'ClassifierBettor',
     'OddsComparisonBettor',
     'backtest',
+    'build_bettor',
     'derive_complementary_events',
     'derive_market_base',
     'find_latest_odds_column',
