+++ src/sportsbet/mcp/_server.py
@@ -13,11 +13,10 @@ import pandas as pd
 from mcp.server.fastmcp import FastMCP
 from sklearn.model_selection import TimeSeriesSplit
 
-from .. import build_bettor, build_dataloader, build_venue
-from .._factory import DEFAULT_KEY_ENV
-from ..dataloaders import load_dataloader
+from ..dataloaders import build_dataloader, load_dataloader
+from ..dataloaders._factory import DEFAULT_KEY_ENV
 from ..evaluation import backtest as run_backtest
-from ..evaluation import load_bettor, save_bettor
+from ..evaluation import build_bettor, load_bettor, save_bettor
 from ..execution import (
     BaseVenue,
     BetIdentity,
@@ -25,6 +24,7 @@ from ..execution import (
     ExposureLimits,
     PlacementIntent,
     PlacementQuote,
+    build_venue,
     value_bet_intents,
 )
 from ..execution import execute as run_execute
