+++ src/sportsbet/sources/_stats/_nba.py
@@ -12,7 +12,7 @@ from typing import Any, ClassVar, Self
 import numpy as np
 import pandas as pd
 
-from ... import ParamGrid
+from ...core import ParamGrid
 from .._base import BaseStatsSource, RawItem, RawPayload
 from .._utils import derive_market_outcomes
